Scaffold the app - ```yo webapp```
One of the advantages of using Yeoman is that it scaffolds your app so that instead of starting from an empty folder, the structure is automatically created and you can start hacking straight away. ```yeo``` is the tool responsible for this part of the process and out of the box it comes bundled with two *generators*. The ```webapp``` generator creates a basic hello-world app containing Modernizr, jQuery, HTML5 Boilerplate and optionally Twitter Bootstrap and Require.js. The other is specifically for Angular, which seems a bit restricted but it looks like this is going to change, [the defaults are being removed and a wide variety made available via npm](https://github.com/yeoman/yo/pull/21).

Running ```yo webapp`` from an empty folder asks whether you'd like to install the optional components. First up Twitter Bootstrap for Sass - what happened to the less version? Turns out [the new version](https://github.com/yeoman/generator-bootstrap/commit/f1cacf0da39b849d427dc122eaa3936169730538) supports less but the currently released version doesn't, I chose not to install it for now. Second up Require.js, again disappointing not to see an alternative for Browserify but not the end of the world.

The final hurdle was a conflict in the ```.gitignore``` file to which I'd already added by editor files but Yeoman was now trying to update with it's set of ignores. Annoyingly I wasn't given the option to merge, I was only asked to overwrite or abort, unsure where abort would leave me I chose overwrite and used git to merge the changes once the process had complete.

At this point I had an app which I could build with ```grunt```, but after looking through the files I couldn't help thinking that hacking them to get them to get to the tool/components setup I wanted might now take longer than just starting from scratch...