```bower install --save bootstrap```
Adding the less version of bootstrap to the project is just a case of running ```bower install --save bootstrap```. ```bower``` is a client-side package manager. ```install --save``` tells bower to fetch the package and save a reference to it in the ```components.json``` file. And finally ```bootstrap``` is the name of the bower package containing the less version of Twitter Bootstrap.

In essense all we've done is cloned Bootstrap into bower ```app/components``` directory, it's up to us to now add it to the build.