```npm install --save-dev grunt-contrib-less```
Where ```yeo``` was responsible for scaffolding the project, ```grunt``` is responsible for actually building it. In order to integrate less in
to the build, we need to edit ```Gruntfile.js``` which is a plain JavaScript file in CommonJS format. Weirdly there are a lot of references to
compass included in the file despite choosing not to install the Sass version of Twitter Boostrap, luckily though it gives us a good place to s
tart integrating the less plugin.

The plugin is installed using ```npm install --save-dev grunt-contrib-less``` and configured to build the ```app/style/main.less``` file to ```.tmp/style/main.css``` (copying the compass plugin). This directory seems to be served up in preference to the app directory when in ```server``` mode and is the place the minifiers look when building in ```dist``` mode. All that's left is to add a sample ```app/style/main.less``` file referencing the main bootstrap less file (```../components/bootstrap/less/bootstrap```), run ```grunt``` and cross my fingers.