```npm install --save-dev grunt-browserify```
Following the same pattern as before, I tried to swap the ```coffee``` plugin for the ```browserify``` one (I know they're not equivalent tools but they're close enough that the config was similar). Unfortunately this is where I ran into bigger and on reflection obvious problems. All of the client packages are managed using ```bower``` which doesn't follow the ```npm``` conventions which ```browserify``` expects.

At this point colour me bored!